# Consolidate and optimize zero functions.