--- a/snacl/snacl.go
+++ b/snacl/snacl.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"runtime/debug"
 
+	"github.com/btcsuite/btcwallet/internal/zero"
 	"github.com/btcsuite/fastsha256"
 	"github.com/btcsuite/golangcrypto/nacl/secretbox"
 	"github.com/btcsuite/golangcrypto/scrypt"
@@ -22,13 +23,6 @@ var (
 	ErrMalformed       = errors.New("malformed data")
 	ErrDecryptFailed   = errors.New("unable to decrypt")
 )
-
-// Zero out a byte slice.
-func zero(b []byte) {
-	for i := range b {
-		b[i] ^= b[i]
-	}
-}
 
 const (
 	// Expose secretbox's Overhead const here for convenience.
@@ -79,7 +73,7 @@ func (ck *CryptoKey) Decrypt(in []byte) ([]byte, error) {
 // rather than waiting until it's reclaimed by the garbage collector.  The
 // key is no longer usable after this call.
 func (ck *CryptoKey) Zero() {
-	zero(ck[:])
+	zero.Bytea32((*[KeySize]byte)(ck))
 }
 
 // GenerateCryptoKey generates a new crypotgraphically random key.
@@ -120,7 +114,7 @@ func (sk *SecretKey) deriveKey(password *[]byte) error {
 		return err
 	}
 	copy(sk.Key[:], key)
-	zero(key)
+	zero.Bytes(key)
 
 	// I'm not a fan of forced garbage collections, but scrypt allocates a
 	// ton of memory and calling it back to back without a GC cycle in