Removes an XXX from the votingpool package.

I didn't touch the zero functions in the keystore package as it is being deprecated and would like to avoid any unnecessary merge conflicts from moving that to the legacy directory.

Zeroing is also performed in hdkeychain, so do we want this here, or somewhere else in our libraries so hdkeychain can also take advantage?  Alternatively, when the btcutil repo is split, we can move hdkeychain to btcwallet, and then update it to use this package.


Benchmarks from a xeon e3 (Bytes is the byte slice zeroing func provided by this package, Bytea* are the fixed-size funcs):

```
BenchmarkXor32  30000000                52.1 ns/op
BenchmarkXor64  20000000                91.5 ns/op
BenchmarkRange32        50000000                31.8 ns/op
BenchmarkRange64        30000000                49.5 ns/op
BenchmarkBytes32        200000000               10.1 ns/op
BenchmarkBytes64        100000000               15.4 ns/op
BenchmarkBytea32        1000000000               2.24 ns/op
BenchmarkBytea64        300000000                4.46 ns/op
```