@jrick 2015-02-04
Updated branch with tests.