@jrick 2015-02-07
Added funcs to zero 32 and 64-byte arrays.  Also moved the zero arrays for Bytes and BigInt inside the function (so they live on the stack and are cleared every time the func is entered).  This shows slower benchmarks due to initializing the same memory over and over again, but should be much faster than a cache miss, and these still benchmark faster than the simple loops which set each byte individually.