@davecgh 2015-03-06
Think this package should have a README.md to be consistent with the others like walletdb/waddrmgr/votingpool/etc?