@jrick 2015-03-06
Depends. Do we expect others to use this package, or should it be an `internal` package?