@davecgh 2015-03-06
It can probably be an internal package.  I like having a README.md for documentation purposes when people are learning how all the code fits together, but if you don't think it's worth it, that's fine.