@jrick 2015-03-06
```Go
+// Package zero contains functions to clear data from byte slices and
+// multi-precision integers.
+package zero
```

Good enough for me :)