@jrick 2015-03-06
Rebased+squashed, and made zero an internal package.